move prosemirror types
