+++ components/common/BoardEditor/focalboard/src/blocks/commentBlock.ts
@@ -1,4 +1,4 @@
-import type { PageContent } from 'models';
+import type { PageContent } from 'lib/prosemirror/interfaces';
 
 import type { Block } from './block';
 import { createBlock } from './block';
