+++ components/common/CharmEditor/CharmEditor.tsx
@@ -32,9 +32,9 @@ import { useSnackbar } from 'hooks/useSnackbar';
 import { useUser } from 'hooks/useUser';
 import log from 'lib/log';
 import type { IPagePermissionFlags } from 'lib/permissions/pages/page-permission-interfaces';
+import type { PageContent } from 'lib/prosemirror/interfaces';
 import { extractDeletedThreadIds } from 'lib/prosemirror/plugins/inlineComments/extractDeletedThreadIds';
 import { setUrlWithoutRerender } from 'lib/utilities/browser';
-import type { PageContent } from 'models';
 
 import * as bulletList from './components/bulletList';
 import Callout, * as callout from './components/callout';
