+++ components/common/CharmEditor/components/nestedPage/components/NestedPage.tsx
@@ -7,7 +7,7 @@ import PageIcon from 'components/common/PageLayout/components/PageIcon';
 import { useCurrentSpace } from 'hooks/useCurrentSpace';
 import { usePages } from 'hooks/usePages';
 import { checkIsContentEmpty } from 'lib/prosemirror/checkIsContentEmpty';
-import type { PageContent } from 'models';
+import type { PageContent } from 'lib/prosemirror/interfaces';
 
 const NestedPageContainer = styled(Link)`
   align-items: center;
