+++ components/common/PageLayout/components/NewPageMenu.tsx
@@ -7,10 +7,10 @@ import Menu from '@mui/material/Menu';
 import MenuItem from '@mui/material/MenuItem';
 import Tooltip from '@mui/material/Tooltip';
 import Typography from '@mui/material/Typography';
+import type { Page } from '@prisma/client';
 import type { MouseEvent } from 'react';
 import { useState, memo } from 'react';
 
-import type { Page } from 'models';
 import { greyColor2 } from 'theme/colors';
 
 import { StyledDatabaseIcon } from './PageIcon';
