+++ models/index.ts
@@ -1,2 +1,1 @@
 export * from './User';
-export * from './Page';
