+++ scripts/updateGalleryImage.ts
@@ -1,6 +1,6 @@
 import { prisma } from 'db';
 import { checkIsContentEmpty } from 'lib/prosemirror/checkIsContentEmpty';
-import { PageContent } from 'models';
+import type { PageContent } from 'lib/prosemirror/interfaces';
 import type { PutObjectCommandInput } from '@aws-sdk/client-s3';
 import { getUserS3FilePath, uploadFileToS3 } from 'lib/aws/uploadToS3Server';
 import { v4 as uuid } from 'uuid';
